% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batplotr-package.R
\docType{package}
\name{batplotr}
\alias{batplotr-package}
\alias{batplotr}
\title{batplotr: Simple BatScope Data Visualization}
\description{
read data from batLogger export and visualize it
}
\section{Data functions}{


     \code{\link{readBatscopeXLSX}} - read .xlsx generated by BatScope export

     \code{\link{sumBatscopeData}} - summarizes Data (from readBatscopeXLSX) and prepares it for plotting
}

\section{Plot functions}{


     \code{\link{night_plot}} - activity over one or more consecutive nights

     \code{\link{period_plot}} - activity over longer periods
}

\section{Accessory Functions}{


     \code{\link{timeOfNight}}

     \code{\link{listHead}}
}

\author{
\strong{Maintainer}: Daniel Angst \email{dcangst@gmail.com}

}
\keyword{BatScope}
\keyword{Data}
\keyword{Simple}
\keyword{batplotr}
\keyword{internal}
\keyword{visualization}
