% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/acc_functions.R
\name{convertToUTC}
\alias{convertToUTC}
\title{Convert to UTC}
\usage{
convertToUTC(date)
}
\arguments{
\item{date}{a POSIXct object, or a vector thereof}
}
\description{
converts time to UTC (to avoid problems with daylight saving time)
}
\seealso{
Other accessory functions: \code{\link{github_update}};
  \code{\link{listHead}}; \code{\link{timeOfNight}}
}

