% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acc_functions.R
\name{getBatLoggerGPS}
\alias{getBatLoggerGPS}
\title{Get BatLogger GPS coordinates}
\usage{
getBatLoggerGPS(GPSCoord, GPSValid)
}
\arguments{
\item{GPSCoord}{GPS coordinates}

\item{GPSValid}{is the value of GPSCoord reliable?}
}
\description{
Get BatLogger GPS coordinates
}
\seealso{
Other accessory functions: \code{\link{convertToUTC}};
  \code{\link{listHead}}; \code{\link{timeOfNight}}
}

