% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_functions.R
\name{github_update}
\alias{github_update}
\title{Check for Package updates on GitHub}
\usage{
github_update(package, repo, ref = "master")
}
\arguments{
\item{package}{the name of the package hosted at \link{repo}}

\item{repo}{Repository address in the format \link{username/repo }}

\item{ref}{Desired git reference. Defaults to "master".}
}
\description{
checks for Package updates
}
\seealso{
Other accessory functions: \code{\link{convertToUTC}},
  \code{\link{listHead}}, \code{\link{timeOfNight}}
}

