% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{night_plot}
\alias{night_plot}
\title{Night Plot}
\usage{
night_plot(plotData, day = unique(plotData$SurveyDate)[1],
  sel_species = "every")
}
\arguments{
\item{plotData}{data generated by \link{\code{sumBatscopeData}}}

\item{day}{day to plot in POSIXct format or "YYYY-MM-DD"}

\item{sel_species}{species_name or vector of species to plot, use "every" to
plot all species individually (differentiated by color)}
}
\value{
a \code{ggplot} object
}
\description{
Plot activity (number of calls) over one night.
}
\seealso{
Other plot functions: \code{\link{year_plot_all}};
  \code{\link{year_plot}}
}

