% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acc_functions.R
\name{nights_SRSS}
\alias{nights_SRSS}
\title{Calculate sunset & sunrise for the night}
\usage{
nights_SRSS(nights)
}
\arguments{
\item{nights}{dates of the nights}
}
\description{
function description
}
\seealso{
Other accessory functions: \code{\link{listHead}};
  \code{\link{timeOfNight}}
}

