% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{periodPlot}
\alias{periodPlot}
\title{Period Plot}
\usage{
periodPlot(plotData, start_date = floor_date(min(plotData$SurveyDate),
  "year"), end_date = ceiling_date(max(plotData$SurveyDate), "year"),
  sel_species = "every", x_limits = NULL, y_limits = NULL)
}
\arguments{
\item{plotData}{data generated by \link{\code{sumBatscopeData}}}

\item{start_date}{in POSIXct format or "YYYY-MM-DD", defaults to the
beginning of the year in which the first data were aquired.}

\item{end_date}{in POSIXct format or "YYYY-MM-DD", defaults to the end
of the year in which the last data were aquired.}

\item{sel_species}{species_name or vector of species to plot, use "every" to
plot all species individually (differentiated by color)}

\item{x_limits}{x-axis limits, NULL (default) or a POSIXct vector of length 2}

\item{y_limits}{y-axis limits, NULL (default) or a numeric vector of length 2}
}
\value{
a \code{ggplot} object
}
\description{
Plot activity (number of calls) of each night of a given period.
}
\seealso{
Other plot functions: \code{\link{nightPlot}}
}

