% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_functions.R
\name{readBatscopeXLSX}
\alias{readBatscopeXLSX}
\title{read Batscope export}
\usage{
readBatscopeXLSX(path, quality_threshold)
}
\arguments{
\item{path}{filename}
}
\description{
reads the \code{xlsx} generated by the batscope export
}

