% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{readBatscopeXLSXmultiple}
\alias{readBatscopeXLSXmultiple}
\title{Read all BatScope Export Files in Folder}
\usage{
readBatscopeXLSXmultiple(path = file.choose(), read_folder = FALSE, ...)
}
\arguments{
\item{path}{path to one file within folder or a vector of path, defaults to choose.file()}

\item{read_folder}{logical, should all .xlsx files in the folder be read?}

\item{...}{additional arguments passed to \code{\link{readBatscopeXLSX}}}
}
\description{
Wrapper for readBatscopeXLSX to read multiple \code{xlsx} files or all files
within a folder
}
\seealso{
Other data functions: 
\code{\link{readBatscopeXLSX}()},
\code{\link{sumBatscopeData}()}
}
\concept{data functions}
