% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_functions.R
\name{sumBatscopeData}
\alias{sumBatscopeData}
\title{Summarize Batscope data_r}
\usage{
sumBatscopeData(data_r, bin_length = 5, nacht_start = 13, nacht_ende = 12)
}
\arguments{
\item{data_r}{data.frame generated with \link{\code{readBatscopeXLSX}}}

\item{nacht_start}{integer, clocktime (hour) when the night starts
(for binning)}

\item{nacht_ende}{integer, clocktime (hour) when the night ends
(for binning)}

\item{bin_width}{length of bins in min}
}
\description{
summarizes the Batscope Data by nights and bins
}
\seealso{
Other data functions: \code{\link{readBatscopeXLSX}}
}

