% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interactive_functions.R
\name{textInput}
\alias{textInput}
\title{textInput}
\usage{
textInput(inputId, label, value = "", ...)
}
\value{
text input
}
\description{
shiny helper function
}
\seealso{
Other interactive functions: \code{\link{shiny_batPlots}}
}

