% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acc_functions.R
\name{timeOfNight}
\alias{timeOfNight}
\title{Time of Night}
\usage{
timeOfNight(date)
}
\arguments{
\item{date}{a POSIXct object, or a vector thereof}
}
\description{
gives back a \code{POSIXct} object with date 1900-01-01 (evening) or
1900-01-02 but with same timestamp for plotting, corrects for daylight saving
time
}
\seealso{
Other accessory functions: \code{\link{convertToUTC}};
  \code{\link{listHead}}
}

