% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{year_plot_all}
\alias{year_plot_all}
\title{Year Plot}
\usage{
year_plot_all(plotData, sunriseData, sunsetData)
}
\arguments{
\item{plotData}{data generated by}
}
\description{
Plot activity (number of calls) over one night for every data point.
}
\seealso{
Other plot functions: \code{\link{night_plot}};
  \code{\link{year_plot}}
}

