% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interactive_functions.R
\name{batPlotR}
\alias{batPlotR}
\title{batPlotR}
\usage{
batPlotR(shiny_options = list(shiny.maxRequestSize = 100 * 1024^2,
  launch.browser = TRUE))
}
\arguments{
\item{shiny_options}{a list of options passed to shinyApp}
}
\value{
launches a WebApp
}
\description{
launches an interactive web app
}

