% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interactive_functions.R
\name{batPlotR}
\alias{batPlotR}
\title{batPlotR}
\usage{
batPlotR(option.list = list(shiny.launch.browser = TRUE, shiny.maxRequestSize
  = 100 * 1024^2))
}
\arguments{
\item{option.list}{a list of global shiny options}
}
\value{
launches a WebApp
}
\description{
launches an interactive web app
}

