% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/batplotr.R
\docType{package}
\name{batplotr}
\alias{batplotr}
\alias{batplotr-package}
\title{rrobot.
A package containing tools for worklist generation for the freedom evo.}
\description{
rrobot.
A package containing tools for worklist generation for the freedom evo.
}
\section{Data functions}{


     \code{\link{readBatscopeXLSX}}

     \code{\link{sumBatscopeData}}
}

\section{Plot functions}{


     \code{\link{night_plot}}

     \code{\link{period_plot}}
}

\section{Accessory Functions}{


     \code{\link{timeOfNight}}

     \code{\link{convertToUTC}}

     \code{\link{listHead}}
}

