% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acc_functions.R
\name{listHead}
\alias{listHead}
\title{Print Head of a list of data.frames}
\usage{
listHead(list, n = "5")
}
\arguments{
\item{list}{a list of \code{data.frame}s}

\item{n}{number of rows to print}
}
\description{
Displays the first \code{n} rows of a data.frame
}
\seealso{
Other accessory functions: \code{\link{nights_SRSS}};
  \code{\link{timeOfNight}}
}

