% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{night_plot}
\alias{night_plot}
\title{Night Plot}
\usage{
night_plot(plotData)
}
\arguments{
\item{plotData}{data generated by}
}
\description{
Plot activity (number of calls) over one night.
}
\seealso{
Other plot functions: \code{\link{year_plot_all}};
  \code{\link{year_plot}}
}

