% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{night_plot}
\alias{night_plot}
\title{Night Plot}
\usage{
night_plot(plotData, day = unique(plotData$SurveyDate)[1],
  sel_species = "every", x_limits = NULL, y_limits = NULL)
}
\arguments{
\item{plotData}{data generated by \link{\code{sumBatscopeData}}}

\item{day}{day to plot in POSIXct format or "YYYY-MM-DD"}

\item{sel_species}{species_name or vector of species to plot, use "every" to
plot all species individually (differentiated by color)}

\item{x_limits}{x-axis limits, NULL (default) or a POSIXct vector of length 2}

\item{y_limits}{y-axis limits, NULL (default) or a numeric vector of length 2}
}
\value{
a \code{ggplot} object
}
\description{
Plot activity (number of calls) over one night.
}
\seealso{
Other plot functions: \code{\link{period_plot}}
}

