% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{nightPlot}
\alias{nightPlot}
\title{Night Plot}
\usage{
nightPlot(plotData, day = min(plotData$survey_date),
  sel_species = "every", x_limits = NULL, y_limits = NULL,
  plot_T = FALSE, plot_T_color = "black", n_ybreaks = 5,
  text_size = 16)
}
\arguments{
\item{plotData}{data generated by \code{\link{sumBatscopeData}}}

\item{day}{day to plot in POSIXct format or "YYYY-MM-DD", defaults to first
date in \code{plotData}}

\item{sel_species}{species name or vector of species to plot, use "every" to 
plot all species individually (differentiated by color)}

\item{x_limits}{x-axis limits, NULL (default) or a POSIXct vector of length 2}

\item{y_limits}{y-axis limits, NULL (default) or a numeric vector of length 2}

\item{plot_T}{logical, should Temperature be plotted?}

\item{plot_T_color}{r color string, see \code{\link{colors}}}

\item{n_ybreaks}{roughly how many tickmarks are plotted on the y-Axis}

\item{text_size}{base text size}
}
\value{
a \code{ggplot} object
}
\description{
Plot activity (number of calls) over one night.
}
\seealso{
Other plot functions: \code{\link{periodPlot}}
}
\concept{plot functions}
