% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_functions.R
\name{readBatscopeXLSX}
\alias{readBatscopeXLSX}
\title{read Batscope export}
\usage{
readBatscopeXLSX(path, quality_threshold = 0.8,
  quality_name = "AutoClass1Qual")
}
\arguments{
\item{path}{filename}

\item{quality_threshold}{sequences with species assignment quality below this
threshold will be discarded.}

\item{quality_name}{what is the name of the column which contains the
relevant quality scores}
}
\description{
reads the \code{xlsx} generated by the batscope export
}

