% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_functions.R
\name{readBatscopeXLSX}
\alias{readBatscopeXLSX}
\title{read Batscope export}
\usage{
readBatscopeXLSX(path = NULL, species_col_name = "AutoClass1",
  quality_col_name = "AutoClass1Qual", quality_threshold = 0.8)
}
\arguments{
\item{path}{filename}

\item{species_col_name}{what is the name of the column which contains the
species name}

\item{quality_col_name}{what is the name of the column which contains the
relevant quality scores}

\item{quality_threshold}{sequences with species assignment quality below this
threshold will be discarded.}
}
\description{
reads the \code{xlsx} generated by the batscope export and brings it into
the right format.
}
\seealso{
Other data functions: \code{\link{sumBatscopeData}}
}

