% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{readBatscopeXLSX}
\alias{readBatscopeXLSX}
\title{read Batscope export}
\usage{
readBatscopeXLSX(
  path = file.choose(),
  batscope_version = "BatScope4",
  species_col_name = "Auto Class 1",
  quality_col_name = "Auto Class 1 Conf",
  quality_threshold = 0.8,
  time_zone = "UTC",
  shiny_progress = FALSE,
  shiny_progress_n = 1
)
}
\arguments{
\item{path}{filename}

\item{species_col_name}{what is the name of the column which contains the
species name}

\item{quality_col_name}{what is the name of the column which contains the
relevant quality scores}

\item{quality_threshold}{sequences with species assignment quality below this
threshold will be discarded.}

\item{time_zone}{time zone of timestamps}

\item{shiny_progress}{display more progress info for shiny}

\item{shiny_progress_n}{fraction of progres bar for multiple files}
}
\description{
reads the \code{xlsx} generated by the batscope export and brings it into
the right format.
}
\seealso{
Other data functions: 
\code{\link{readBatscopeXLSXmultiple}()},
\code{\link{sumBatscopeData}()}
}
\concept{data functions}
