% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_functions.R
\name{shiny_batPlots}
\alias{shiny_batPlots}
\title{batPlotR}
\usage{
shiny_batPlots(option.list = list(shiny.launch.browser = TRUE,
  shiny.maxRequestSize = 100 * 1024^2, encoding = "UTF-8"))
}
\arguments{
\item{option.list}{a list of global shiny options}
}
\value{
launches a WebApp
}
\description{
launches an interactive web app
}

