% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_functions.R
\name{sumBatscopeData}
\alias{sumBatscopeData}
\title{Summarize Batscope data_r}
\usage{
sumBatscopeData(data_r, bin_length = 5, nacht_start = 13, nacht_ende = 12,
  lat = NULL, long = NULL, progress = "text")
}
\arguments{
\item{data_r}{data.frame generated with \link{\code{readBatscopeXLSX}}}

\item{nacht_start}{integer, clocktime (hour) when the night starts
(for binning)}

\item{nacht_ende}{integer, clocktime (hour) when the night ends
(for binning)}

\item{lat}{vector of GPS latitudes for the stations
(recycled if needed). If NULL (default) the logged GPS data will be used
(averaged for each station)}

\item{long}{vector of GPS longitudes for the stations
(recycled if needed). If NULL (default) the logged GPS data will be used
(averaged for each station)}

\item{progress}{name of the progress bar to use, see
\link{\code{plyr::create_progress_bar}}}

\item{bin_width}{length of bins in min}
}
\description{
summarizes the Batscope Data by nights and bins
}
\seealso{
Other data functions: \code{\link{readBatscopeXLSX}}
}

