% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{sumBatscopeData}
\alias{sumBatscopeData}
\title{Summarize Batscope data_r}
\usage{
sumBatscopeData(
  data_r,
  bin_length = 5,
  lat = NULL,
  long = NULL,
  progress = "text",
  shiny_progress = FALSE
)
}
\arguments{
\item{data_r}{data.frame generated with \code{\link{readBatscopeXLSX}}}

\item{lat}{vector of GPS latitudes for the stations
(recycled if needed). If NULL (default) the logged GPS data will be used 
(averaged for each station)}

\item{long}{vector of GPS longitudes for the stations
(recycled if needed). If NULL (default) the logged GPS data will be used 
(averaged for each station)}

\item{progress}{name of the progress bar to use, see 
\code{\link[plyr]{create_progress_bar}}}

\item{shiny_progress}{display more progress info for shiny}

\item{bin_width}{length of bins in min}
}
\description{
summarizes the Batscope Data by nights and bins
}
\seealso{
Other data functions: 
\code{\link{readBatscopeXLSX_multiple}()},
\code{\link{readBatscopeXLSX}()}
}
\concept{data functions}
