% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acc_functions.R
\name{timeOfNight}
\alias{timeOfNight}
\title{Time of Night}
\usage{
timeOfNight(df)
}
\arguments{
\item{date}{a POSIXct object}
}
\description{
gives back a \code{POSIXct} object with date 1900-01-01 (evening) or
1900-01-02 but with same timestamp for plotting
}
\seealso{
Other accessory functions: \code{\link{listHead}}
}

